Themes: Watch all app crashes for problematic themes
Only watching for exceptions that occur during an apps launch may
not be sufficient to catch possible bad themes causing issues.
This patch monitors all app crashes and broadcasts that there was
a crash if either Resources.NotFoundException or InflateException
are thrown.  The remaining logic is the same as when an app launch
failure occurred.

We also display a notification to the user so that they know why
their theme was reset to the system theme.

Change-Id: I0761d02587b5b81deee4a31a89f515dbc7cc5fe6

Conflicts:
	core/res/res/values/cm_symbols.xml